add async update_article() to fix coar resubmission
Details: send_acknowlege_and_reject() needs to update
the 'article' db item for coar resubmission use-cases,
but is called in the context of a 'before update' db trigger
already, which prevents effective db update - likely blocked
or ignored by web2py. So we thead the db update to aleviate
the constraint.